Improve token handling (#30)
Closes #29

This PR deprecates the `createUnauthenticatedAuth` method in the
`@octokit/auth-unauthenticated` library.

In environments such as the GitHub Enterprise Server (GHES), `octokit`
instances will be set to `null` if the input `token` is not explicitly
provided. To improve environmental resilience and functional stability,
the `downloadTool` method in the `@actions/tool-cache` library will be
utilized when `octokit` is `null` to securely call the GitHub API over
the HTTPS protocol to ensure accurate access to project release
information.

The goal of this adjustment is to ensure smooth access to the necessary
release information in different scenarios, especially in enterprise
environments, so that our tool can efficiently support subsequent
processes and improve overall compatibility and security.